pythonpostgres/requirements.txt
astroid==2.4.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pylint==2.6.0
redis==3.5.3
six==1.15.0
toml==0.10.2
wrapt==1.12.1
